Machine-Learning_API/requirements.txt
tensorflow
keras_preprocessing
pandas
numpy
fastapi
uvicorn
pillow